+++ app.vue
@@ -2,6 +2,7 @@
   <div id="app" class="bg-gradient-to-r from-purple-500 to-pink-500 text-dark dark:text-light">
     <HeaderAppHeader :section-list="sectionList" />
     <NuxtPage :selected-language-data="selectedLanguageData" />
+    <AppFooter />
   </div>
 </template>