+++ components/AppFooter.vue
@@ -1,0 +1,13 @@
+<template>
+  <footer class="flex items-center justify-center w-full gap-6 p-6 mx-auto bg-primary-light dark:bg-primary-dark lg:py-8 shadow-top">
+    <span class="text-4xl"> > </span>
+    <Icon
+      name="MainLogo"
+      width="258.808"
+      height="54.308"
+      class="h-12 fill-current"
+      alt="Main logo of the app"
+    />
+    <span class="text-4xl"> &lt; </span>
+  </footer>
+</template>