+++ components/header/AppHeader.vue
@@ -11,34 +11,21 @@
   <!-- Header of the app -->
   <header
     id="header"
-    class="fixed top-0 z-50 flex flex-col w-full gap-6 px-6 py-4 transition-all bg-transparent shadow-lg md:px-8 backdrop-blur-lg"
+    class="fixed top-0 z-50 flex items-center justify-between w-full gap-6 px-6 py-4 transition-all shadow-lg h-min bg-primary-light dark:bg-primary-dark md:px-8 scroll-reduce"
   >
-    <nav class="container flex items-center content-between justify-between h-auto gap-4 mx-auto">
-      <div class="flex justify-start w-36 lg:w-48">
-        <Icon
-          name="MainLogo"
-          width="258.808"
-          height="54.308"
-          class="w-full h-full fill-current"
-          alt="Main logo of the app"
-        />
-      </div>
-      <div class="flex justify-center">
-        <Icon
-          name="ph:list-bold"
-          class="w-5 h-5 cursor-pointer fill-current md:w-6 md:h-6 hover:text-purple-500"
-          :class="{ 'text-purple-700' : menuOpen }"
-          role="button"
-          aria-label="Toggle menu"
-          :aria-expanded="menuOpen.toString()"
-          @click="toggleMenuDropdown"
-        />
-      </div>
-      <div class="flex flex-row items-center justify-end gap-3 md:gap-6">
-        <HeaderThemeSwitch aria-label="Switch theme" />
-        <HeaderLanguageSelector aria-label="Select language" />
-      </div>
-    </nav>
+    <Icon
+      name="ic:round-list"
+      class="cursor-pointer fill-current w-7 h-7 md:w-8 md:h-8 hover:text-purple-500"
+      :class="{ 'text-purple-700' : menuOpen }"
+      role="button"
+      aria-label="Toggle menu"
+      :aria-expanded="menuOpen.toString()"
+      @click="toggleMenuDropdown"
+    />
+    <div class="flex flex-row items-center gap-3 md:gap-6">
+      <HeaderThemeSwitch aria-label="Switch theme" />
+      <HeaderLanguageSelector aria-label="Select language" />
+    </div>
   </header>
 </template>
 
@@ -57,3 +44,24 @@ const toggleMenuDropdown = () => {
   menuOpen.value = !menuOpen.value
 }
 </script>
+
+<style scoped>
+@keyframes reduce {
+  to {
+    background: transparent;
+    backdrop-filter: blur(16px);
+    padding-left: 3rem;
+    padding-right: 3rem;
+    padding-top: 0.5rem;
+    padding-bottom: 0.5rem;
+  }
+
+}
+
+.scroll-reduce {
+  animation: linear reduce both;
+  animation-range: 0 150vh;
+  animation-timeline: scroll(root);
+}
+
+</style>